Refactor MainWorker imports to use renamed worker file
